@@ -10,27 +10,27 @@
   "author": "Final Child",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^3.0.0",
-    "koa": "^2.5.2",
+    "bcrypt": "^3.0.6",
+    "koa": "^2.7.0",
     "koa-bodyparser": "^4.2.1",
     "koa-ip-filter": "^3.0.0",
     "koa-router": "^7.4.0",
     "koa-send": "^5.0.0",
     "koa-static": "^5.0.0",
     "mz": "^2.7.0",
-    "sqlite3": "^4.0.2",
+    "sqlite3": "^4.0.9",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "@types/bcrypt": "^2.0.0",
-    "@types/koa": "^2.0.46",
+    "@types/bcrypt": "^3.0.0",
+    "@types/koa": "^2.0.49",
     "@types/koa-bodyparser": "^5.0.1",
-    "@types/koa-router": "^7.0.30",
-    "@types/koa-static": "^4.0.0",
+    "@types/koa-router": "^7.0.42",
+    "@types/koa-static": "^4.0.1",
     "@types/mz": "0.0.32",
-    "@types/node": "^10.5.2",
-    "@types/sqlite3": "^3.1.3",
-    "ts-node": "^7.0.0",
-    "typescript": "^2.9.2"
+    "@types/node": "^12.6.6",
+    "@types/sqlite3": "^3.1.5",
+    "ts-node": "^8.3.0",
+    "typescript": "^3.5.3"
   }
 }
