@@ -12,49 +12,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.9",
-    "@angular/cdk": "^6.4.0",
-    "@angular/common": "^6.0.9",
-    "@angular/compiler": "^6.0.9",
-    "@angular/core": "^6.0.9",
-    "@angular/forms": "^6.0.9",
-    "@angular/http": "^6.0.9",
-    "@angular/material": "6.4.0",
-    "@angular/platform-browser": "^6.0.9",
-    "@angular/platform-browser-dynamic": "^6.0.9",
-    "@angular/router": "^6.0.9",
-    "@swimlane/ngx-charts": "^9.0.0",
+    "@angular/animations": "^8.1.1",
+    "@angular/cdk": "^8.1.1",
+    "@angular/common": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@angular/forms": "^8.1.1",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "8.1.1",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.1.1",
+    "@swimlane/ngx-charts": "^12.0.1",
     "checkdigit": "^1.1.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.5.7",
-    "d3": "^5.5.0",
-    "file-saver": "^1.3.8",
+    "core-js": "^3.1.4",
+    "d3": "^5.9.7",
+    "file-saver": "^2.0.2",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.2.2",
-    "xlsx-populate": "^1.17.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.2",
+    "xlsx-populate": "^1.19.1",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "6.0.8",
-    "@angular/compiler-cli": "^6.0.9",
-    "@angular/language-service": "^6.0.9",
-    "@types/file-saver": "1.3.0",
-    "@types/jasmine": "^2.8.8",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^10.5.2",
-    "codelyzer": "~4.4.2",
-    "jasmine-core": "~3.1.0",
+    "@angular-devkit/build-angular": "~0.801.1",
+    "@angular/cli": "8.1.1",
+    "@angular/compiler-cli": "^8.1.1",
+    "@angular/language-service": "^8.1.1",
+    "@types/file-saver": "2.0.1",
+    "@types/jasmine": "^3.3.13",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "^12.6.6",
+    "codelyzer": "~5.1.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.4",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^2.0.1",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^1.2.0",
-    "protractor": "^5.3.2",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "^2.7.2"
+    "karma": "^4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "^5.4.2",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "<3.5"
   }
 }
