@@ -3,7 +3,7 @@ import CandidateNamesCache from './common/CandidateNamesCache';
 let candidateNamesCache: CandidateNamesCache;
 
 export function getCache(): CandidateNamesCache {
-    if (candidateNamesCache === undefined) {
+    if (candidateNamesCache === undefined || candidateNamesCache === null) {
         throw 'candidateNamesCache is undefined!';
     }
     return candidateNamesCache;
