@@ -69,12 +69,18 @@ export async function tryToSetVoted(key: number): Promise<void> {
 /**
  * 투표
  * @param key 투표자 키
+ * @param grade 투표자 학년
  * @param candidateName1M 1학년 남자 후보자 이름 (falsy인 경우 무시)
  * @param candidateName1F 1학년 여자 후보자 이름 (falsy인 경우 무시)
  * @param candidateName2 2학년 후보자 이름 (falsy인 경우 무시)
  */
-export async function vote(key: number, candidateName1M: string, candidateName1F: string, candidateName2: string): Promise<void> {
+export async function vote(key: number, grade: number, candidateName1M: string, candidateName1F: string, candidateName2: string): Promise<void> {
     await new Promise((resolve, reject) => {
+        let gradeVotesColumn = '';
+        if (grade === 1) gradeVotesColumn = 'first_grade_votes';
+        else if (grade === 2) gradeVotesColumn = 'second_grade_votes';
+        else if (grade === 3) gradeVotesColumn = 'third_grade_votes';
+        else throw 'Invalid grade!';
         assertValidKey(key);
         tryToSetVoted(key)
             .then(() => {
@@ -82,17 +88,17 @@ export async function vote(key: number, candidateName1M: string, candidateName1F
                     db.run('BEGIN TRANSACTION');
                     if (candidateName1M) {
                         db.run(`UPDATE candidate1M
-                        SET votes = votes + 1
+                        SET votes = votes + 1, ${gradeVotesColumn} = ${gradeVotesColumn} + 1
                         WHERE name = ?`, [candidateName1M]);
                     }
                     if (candidateName1F) {
                         db.run(`UPDATE candidate1F
-                        SET votes = votes + 1
+                        SET votes = votes + 1, ${gradeVotesColumn} = ${gradeVotesColumn} + 1
                         WHERE name = ?`, [candidateName1F]);
                     }
                     if (candidateName2) {
                         db.run(`UPDATE candidate2
-                        SET votes = votes + 1
+                        SET votes = votes + 1, ${gradeVotesColumn} = ${gradeVotesColumn} + 1
                         WHERE name = ?`, [candidateName2]);
                     }
                     db.run('COMMIT', [], (err) => {
@@ -111,7 +117,7 @@ export function getCandidate1M(candidateName: string) {
     return new Promise((resolve, reject) => {
         if (candidateName) {
             db.serialize(() => {
-                db.get('SELECT * FROM candidate1M WHERE name = ?', [candidateName], (err, row) => {
+                db.get('SELECT name, votes, first_grade_votes as firstGradeVotes, second_grade_votes as secondGradeVotes, third_grade_votes as thirdGradeVotes FROM candidate1M WHERE name = ?', [candidateName], (err, row) => {
                     if (err) {
                         reject(err);
                         return;
@@ -130,7 +136,7 @@ export function getCandidate1F(candidateName: string) {
     return new Promise((resolve, reject) => {
         if (candidateName) {
             db.serialize(() => {
-                db.get('SELECT * FROM candidate1F WHERE name = ?', [candidateName], (err, row) => {
+                db.get('SELECT name, votes, first_grade_votes as firstGradeVotes, second_grade_votes as secondGradeVotes, third_grade_votes as thirdGradeVotes FROM candidate1F WHERE name = ?', [candidateName], (err, row) => {
                     if (err) {
                         reject(err);
                         return;
@@ -149,7 +155,7 @@ export function getCandidate2(candidateName: string) {
     return new Promise((resolve, reject) => {
         if (candidateName) {
             db.serialize(() => {
-                db.get('SELECT * FROM candidate2 WHERE name = ?', [candidateName], (err, row) => {
+                db.get('SELECT name, votes, first_grade_votes as firstGradeVotes, second_grade_votes as secondGradeVotes, third_grade_votes as thirdGradeVotes FROM candidate2 WHERE name = ?', [candidateName], (err, row) => {
                     if (err) {
                         reject(err);
                         return;
@@ -167,7 +173,7 @@ export function getCandidate2(candidateName: string) {
 export async function getCandidates1M(): Promise<Array<Candidate>> {
     return await new Promise<Array<Candidate>>((resolve, reject) => {
         db.serialize(() => {
-            db.all('SELECT * FROM candidate1M', [], (err, rows) => {
+            db.all('SELECT name, votes, first_grade_votes as firstGradeVotes, second_grade_votes as secondGradeVotes, third_grade_votes as thirdGradeVotes FROM candidate1M', [], (err, rows) => {
                 if (err) {
                     reject(err);
                     return;
@@ -181,7 +187,7 @@ export async function getCandidates1M(): Promise<Array<Candidate>> {
 export async function getCandidates1F(): Promise<Array<Candidate>> {
     return await new Promise<Array<Candidate>>((resolve, reject) => {
         db.serialize(() => {
-            db.all('SELECT * FROM candidate1F', [], (err, rows) => {
+            db.all('SELECT name, votes, first_grade_votes as firstGradeVotes, second_grade_votes as secondGradeVotes, third_grade_votes as thirdGradeVotes FROM candidate1F', [], (err, rows) => {
                 if (err) {
                     reject(err);
                     return;
@@ -195,7 +201,7 @@ export async function getCandidates1F(): Promise<Array<Candidate>> {
 export async function getCandidates2(): Promise<Array<Candidate>> {
     return await new Promise<Array<Candidate>>((resolve, reject) => {
         db.serialize(() => {
-            db.all('SELECT * FROM candidate2', [], (err, rows) => {
+            db.all('SELECT name, votes, first_grade_votes as firstGradeVotes, second_grade_votes as secondGradeVotes, third_grade_votes as thirdGradeVotes FROM candidate2', [], (err, rows) => {
                 if (err) {
                     reject(err);
                     return;
@@ -219,23 +225,33 @@ export async function getCandidates(): Promise<Candidates> {
 }
 
 export async function setCandidates(candidates: Candidates): Promise<Candidates> {
-    const statement1M = 'INSERT INTO candidate1M (name, votes) VALUES ' + new Array(candidates.candidates1M.length).fill('(?, ?)').join(', ');
+    const statement1M = 'INSERT INTO candidate1M (name, votes, first_grade_votes, second_grade_votes, third_grade_votes) VALUES ' + new Array(candidates.candidates1M.length).fill('(?, ?, ?, ?, ?)').join(', ');
     const placeholder1M = [];
     candidates.candidates1M.forEach(candidate => {
         placeholder1M.push(candidate.name);
         placeholder1M.push(candidate.votes);
+        placeholder1M.push(candidate.firstGradeVotes);
+        placeholder1M.push(candidate.secondGradeVotes);
+        placeholder1M.push(candidate.thirdGradeVotes);
+
     });
-    const statement1F = 'INSERT INTO candidate1F (name, votes) VALUES ' + new Array(candidates.candidates1F.length).fill('(?, ?)').join(', ');
+    const statement1F = 'INSERT INTO candidate1F (name, votes, first_grade_votes, second_grade_votes, third_grade_votes) VALUES ' + new Array(candidates.candidates1F.length).fill('(?, ?, ?, ?, ?)').join(', ');
     const placeholder1F = [];
     candidates.candidates1F.forEach(candidate => {
         placeholder1F.push(candidate.name);
         placeholder1F.push(candidate.votes);
+        placeholder1F.push(candidate.firstGradeVotes);
+        placeholder1F.push(candidate.secondGradeVotes);
+        placeholder1F.push(candidate.thirdGradeVotes);
     });
-    const statement2 = 'INSERT INTO candidate2 (name, votes) VALUES ' + new Array(candidates.candidates2.length).fill('(?, ?)').join(', ');
+    const statement2 = 'INSERT INTO candidate2 (name, votes, first_grade_votes, second_grade_votes, third_grade_votes) VALUES ' + new Array(candidates.candidates2.length).fill('(?, ?, ?, ?, ?)').join(', ');
     const placeholder2 = [];
     candidates.candidates2.forEach(candidate => {
         placeholder2.push(candidate.name);
         placeholder2.push(candidate.votes);
+        placeholder2.push(candidate.firstGradeVotes);
+        placeholder2.push(candidate.secondGradeVotes);
+        placeholder2.push(candidate.thirdGradeVotes);
     });
 
     await new Promise<Candidates>((resolve, reject) => {
@@ -268,6 +284,7 @@ export async function setCandidates(candidates: Candidates): Promise<Candidates>
 export async function getCandidateNames(): Promise<CandidateNames> {
     const candidates = await getCandidates();
     return {
+        pollName: await getPollName(),
         candidateNames1M: candidates.candidates1M.map(candidate => candidate.name),
         candidateNames1F: candidates.candidates1F.map(candidate => candidate.name),
         candidateNames2: candidates.candidates2.map(candidate => candidate.name)
@@ -275,22 +292,32 @@ export async function getCandidateNames(): Promise<CandidateNames> {
 }
 
 export async function setCandidateNames(candidateNames: CandidateNames): Promise<void> {
-    const statement1M = 'INSERT INTO candidate1M (name, votes) VALUES ' + new Array(candidateNames.candidateNames1M.length).fill('(?, ?)').join(', ');
+    await setPollName(candidateNames.pollName);
+    const statement1M = 'INSERT INTO candidate1M (name, votes, first_grade_votes, second_grade_votes, third_grade_votes) VALUES ' + new Array(candidateNames.candidateNames1M.length).fill('(?, ?, ?, ?, ?)').join(', ');
     const placeholder1M = [];
     candidateNames.candidateNames1M.forEach(candidateName => {
         placeholder1M.push(candidateName);
         placeholder1M.push(0);
+        placeholder1M.push(0);
+        placeholder1M.push(0);
+        placeholder1M.push(0);
     });
-    const statement1F = 'INSERT INTO candidate1F (name, votes) VALUES ' + new Array(candidateNames.candidateNames1F.length).fill('(?, ?)').join(', ');
+    const statement1F = 'INSERT INTO candidate1F (name, votes, first_grade_votes, second_grade_votes, third_grade_votes) VALUES ' + new Array(candidateNames.candidateNames1F.length).fill('(?, ?, ?, ?, ?)').join(', ');
     const placeholder1F = [];
     candidateNames.candidateNames1F.forEach(candidateName => {
         placeholder1F.push(candidateName);
         placeholder1F.push(0);
+        placeholder1F.push(0);
+        placeholder1F.push(0);
+        placeholder1F.push(0);
     });
-    const statement2 = 'INSERT INTO candidate2 (name, votes) VALUES ' + new Array(candidateNames.candidateNames2.length).fill('(?, ?)').join(', ');
+    const statement2 = 'INSERT INTO candidate2 (name, votes, first_grade_votes, second_grade_votes, third_grade_votes) VALUES ' + new Array(candidateNames.candidateNames2.length).fill('(?, ?, ?, ?, ?)').join(', ');
     const placeholder2 = [];
     candidateNames.candidateNames2.forEach(candidateName => {
         placeholder2.push(candidateName);
+        placeholder2.push(0);
+        placeholder2.push(0);
+        placeholder2.push(0);
         placeholder2.push(0);
     });
     await new Promise((resolve, reject) => {
