@@ -1,13 +1,13 @@
-import StudentsRequest from '../../common/request/admin/StudentRequest';
+import StudentsRequest from '../../common/request/admin/StudentsRequest';
 import Router = require('koa-router');
 import {assertValidAdminPassword} from '../../common/util';
 import {compareAdminPassword, getStudents} from '../../database';
 
-const router = new Router({prefix: '/api/admin/students'});
+const router = new Router({prefix: '/students'});
 export default router;
 
 router.post('/', async (ctx, next) => {
-    const request: StudentsRequest = ctx.request.body;
+    const request = <StudentsRequest>ctx.request.body;
 
     try {
         assertValidAdminPassword(request.adminPassword);
