@@ -3,21 +3,32 @@ import * as Mustache from "mustache";
 const keyInput = document.getElementById('login-key');
 const keyError = document.getElementById('key-error');
 const keyInputContainer = document.getElementById('key-input-container');
-keyInput.addEventListener('input', async function (e) {
-    if (!keyInput.value || keyInput.value.length !== 8) {
+const studentNumberInput = document.getElementById('login-student-number');
+const studentNumberError = document.getElementById('student-number-error');
+const studentNumberInputContainer = document.getElementById('student-number-input-container');
+
+keyInput.addEventListener('input', onInput);
+studentNumberInput.addEventListener('input', onInput);
+
+async function onInput() {
+    if (!keyInput.value || keyInput.value.length !== 8 || !studentNumberInput.value || studentNumberInput.value.length !== 5) {
         return;
     }
     if (!checkKeyString()) {
         keyInput.focus();
         return;
     }
+    if (!checkStudentNumberString()) {
+        studentNumberInput.focus();
+        return;
+    }
 
     const key = Math.floor(parseInt(keyInput.value, 10) / 10);
-    await login(key);
-});
+    const studentNumber = parseInt(studentNumberInput.value, 10);
+    await login(key, studentNumber);
+}
 
 function checkKeyString() {
-    const keyString = keyInput.value;
     if (!keyInput.value || keyInput.value.length !== 8) {
         keyInput.setCustomValidity('8자리의 키를 입력해 주세요!');
         keyError.textContent = '8자리의 키를 입력해 주세요!';
@@ -34,17 +45,35 @@ function checkKeyString() {
     return true;
 }
 
+function checkStudentNumberString() {
+    if (!studentNumberInput.value || studentNumberInput.value.length !== 5) {
+        studentNumberInput.setCustomValidity('5자리의 학번을 입력해 주세요!');
+        studentNumberError.textContent = '5자리의 학번을 입력해 주세요!';
+        studentNumberInputContainer.classList.add('is-invalid');
+        return false;
+    }
+    const studentNumber = parseInt(studentNumberInput.value);
+    if (!Number.isSafeInteger(studentNumber) || studentNumber < 10000 || studentNumber >= 40000) {
+        studentNumberInput.setCustomValidity('학번이 잘못되었습니다!');
+        keyError.textContent = '학번이 잘못되었습니다!';
+        keyInputContainer.classList.add('is-invalid');
+        return false;
+    }
+    return true;
+}
+
 function isValidKeyWithCheckDigit(key) {
     return Number.isSafeInteger(key) && key > 0 && key < 100000000 && isValidMod10(key);
 }
 
-async function login(key) {
+async function login(key, studentNumber) {
     const cache = JSON.parse(sessionStorage.getItem('cache'));
-    const request = new Request('/api/login',  {
+    const request = new Request('/api/login', {
         method: 'POST',
         body: JSON.stringify({
-            key: key,
-            candidateCacheId: (cache && cache.candidatesCacheId) ? cache.candidatesCacheId : undefined
+            key,
+            studentNumber,
+            candidatesCacheId: (cache !== null && cache.candidatesCacheId !== null) ? cache.candidatesCacheId : null
         }),
         headers: new Headers({
             'content-type': 'application/json'
@@ -60,13 +89,15 @@ async function login(key) {
     }
     sessionStorage.setItem('key', key.toString());
     sessionStorage.setItem('grade', result.grade.toString());
+    sessionStorage.setItem('studentNumber', studentNumber.toString());
     if (result.cache) {
         sessionStorage.setItem('cache', JSON.stringify(result.cache));
     }
+    const newCache = JSON.parse(sessionStorage.getItem('cache'));
 
-    sessionStorage.setItem('rendered', await loadTemplate(result.cache, result.grade));
+    sessionStorage.setItem('rendered', await loadTemplate(newCache, result.grade));
 
-    location = 'vote.html';
+    location.assign('vote.html');
 }
 
 function isValidMod10(number) {
@@ -94,7 +125,7 @@ async function loadTemplate(cache, grade) {
     const candidateNames1M = cache.candidateNames.candidateNames1M;
     const candidateNames1F = cache.candidateNames.candidateNames1F;
     const candidateNames2 = cache.candidateNames.candidateNames2;
-    const pollNameEscaped = encodeURIComponent(cache.pollName);
+    const pollNameEscaped = encodeURIComponent(cache.candidateNames.pollName);
     const abstention1M = candidateNames1M.includes('기권');
     const abstention1F = candidateNames1F.includes('기권');
     const abstention2 = candidateNames2.includes('기권');
